[cognito] Cache JWKs to prevent rate limit errors from Cognito
- The PyJWKClient needs to be persisted for the caching logic to work,
  otherwise it'll just fetch the key each time it's used
